refactor(tests): reorganize assessment endpoint tests into separate file
- Deleted the existing readme-endpoints test file to streamline test organization.
- Created a new assessment-endpoints test file to specifically handle assessment-related API tests.
- Improved modularity and maintainability of test code by separating concerns for better clarity.
