--- backend/tests/e2e/prod/readme-endpoints.test.js
@@ -1,185 +1,0 @@
-import { describe, test, expect } from 'vitest';
-import fetch from 'node-fetch';
-
-// Define the base API URL for production tests
-const API_URL = process.env.VERCEL_URL 
-  ? `https://${process.env.VERCEL_URL}`
-  : 'https://dottie-api-zeta.vercel.app';
-
-// Test timeout - set to 30 seconds for potential slow responses in production
-const TEST_TIMEOUT = 15000;
-
-// Helper function to create a mock token for testing
-const createMockToken = () => {
-  const header = Buffer.from(JSON.stringify({ alg: 'HS256', typ: 'JWT' })).toString('base64');
-  const payload = Buffer.from(JSON.stringify({ 
-    id: `test-user-${Date.now()}`, 
-    email: `test_${Date.now()}@example.com`,
-    exp: Math.floor(Date.now() / 1000) + 3600
-  })).toString('base64');
-  const signature = 'mock_signature'; // This won't validate, but it's for structure testing
-  return `${header}.${payload}.${signature}`;
-};
-
-// Generate a test token
-const testToken = createMockToken();
-const testUserId = `test-user-${Date.now()}`;
-
-describe("README API Endpoints Tests - Production", () => {
-  // Test Group 1: Testing basic endpoints
-  describe("Testing Basic Endpoints", () => {
-    // Test the hello endpoint
-    test("GET /api/hello - should return Hello World message", async () => {
-      const response = await fetch(`${API_URL}/api/hello`);
-      expect(response.status).toBe(200);
-      
-      const data = await response.json();
-      expect(data).toHaveProperty("message");
-      expect(data.message).toBe("Hello World from Dottie API!");
-    }, TEST_TIMEOUT);
-
-    // Test the db-status endpoint - may return 401 in production
-    test("GET /api/db-status - should check database status", async () => {
-      console.log('Testing database status endpoint - this may time out...');
-      const response = await fetch(`${API_URL}/api/db-status`);
-      
-      // In production, this endpoint may require auth
-      if (response.status === 200) {
-        const data = await response.json();
-        expect(data).toHaveProperty("status");
-        console.log(`Database status: ${data.status}`);
-      } else if (response.status === 401) {
-        console.log('Database status endpoint requires authentication');
-      } else if (response.status === 504) {
-        console.error('DATABASE ENDPOINT TIMED OUT (504) - This is now treated as a failure');
-      } else {
-        console.log(`Unexpected status from database endpoint: ${response.status}`);
-      }
-      
-      // Don't accept 504 timeout as valid response
-      expect([200, 401]).toContain(response.status);
-    }, TEST_TIMEOUT);
-  });
-
-  // Test Group 2: User Authentication Endpoints
-  describe("User Authentication Endpoints", () => {
-    // Test user signup - may be restricted in production
-    test("POST /api/auth/signup - should attempt to create a new user", async () => {
-      console.log('Testing signup endpoint - this may time out...');
-      const userData = {
-        username: `testuser_${Date.now()}`,
-        email: `test_${Date.now()}@example.com`,
-        password: "Password123!",
-        age: "18_24"
-      };
-
-      const response = await fetch(`${API_URL}/api/auth/signup`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(userData)
-      });
-      
-      if (response.status === 504) {
-        console.error('SIGNUP ENDPOINT TIMED OUT (504) - This is now treated as a failure');
-      }
-      
-      // In production, this might be locked down or require special permissions
-      // Just check that the endpoint exists and responds
-      console.log(`Signup endpoint status: ${response.status}`);
-      expect([201, 400, 401, 403, 404, 500]).toContain(response.status);
-    }, TEST_TIMEOUT);
-    
-    // Test user login
-    test("POST /api/auth/login - should attempt to authenticate user", async () => {
-      console.log('Testing login endpoint - this may time out...');
-      const loginData = {
-        email: `test_${Date.now()}@example.com`,
-        password: "Password123!"
-      };
-
-      const response = await fetch(`${API_URL}/api/auth/login`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(loginData)
-      });
-      
-      if (response.status === 504) {
-        console.error('LOGIN ENDPOINT TIMED OUT (504) - This is now treated as a failure');
-      }
-      
-      // In production, with no valid credentials, we expect unauthorized
-      console.log(`Login endpoint status: ${response.status}`);
-      expect([200, 401, 403, 404, 500]).toContain(response.status);
-    }, TEST_TIMEOUT);
-  });
-
-  // Test Group 3: Protected Endpoints
-  describe("Protected Endpoints", () => {
-    // Test getting all users (protected)
-    test("GET /api/auth/users - should check authenticated endpoint", async () => {
-      const response = await fetch(`${API_URL}/api/auth/users`, {
-        headers: { "Authorization": `Bearer ${testToken}` }
-      });
-      
-      // We expect 401 since our token is invalid, but endpoint should exist
-      console.log(`Get users endpoint status: ${response.status}`);
-      expect([200, 401, 403, 404]).toContain(response.status);
-    }, TEST_TIMEOUT);
-    
-    // Test getting a specific user by ID (protected)
-    test("GET /api/auth/users/:id - should check specific user endpoint", async () => {
-      const response = await fetch(`${API_URL}/api/auth/users/${testUserId}`, {
-        headers: { "Authorization": `Bearer ${testToken}` }
-      });
-      
-      // We expect 401 since our token is invalid, but endpoint should exist
-      console.log(`Get user by ID endpoint status: ${response.status}`);
-      expect([200, 401, 403, 404]).toContain(response.status);
-    }, TEST_TIMEOUT);
-  });
-
-  // Test Group 4: Assessment Endpoints
-  describe("Assessment Endpoints", () => {
-    // Test submitting assessment (protected)
-    test("POST /api/assessment/send - should check assessment submission", async () => {
-      const assessmentData = {
-        userId: testUserId,
-        assessmentData: {
-          age: "18_24",
-          cycleLength: "26_30",
-          periodDuration: "4_5",
-          flowHeaviness: "moderate",
-          painLevel: "moderate",
-          symptoms: {
-            physical: ["Bloating", "Headaches"],
-            emotional: ["Mood swings"]
-          }
-        }
-      };
-
-      const response = await fetch(`${API_URL}/api/assessment/send`, {
-        method: 'POST',
-        headers: { 
-          'Content-Type': 'application/json',
-          'Authorization': `Bearer ${testToken}`
-        },
-        body: JSON.stringify(assessmentData)
-      });
-      
-      // We expect 401 since our token is invalid, but endpoint should exist
-      console.log(`Assessment submission endpoint status: ${response.status}`);
-      expect([201, 400, 401, 403, 404, 500]).toContain(response.status);
-    }, TEST_TIMEOUT);
-    
-    // Test getting assessments list (protected)
-    test("GET /api/assessment/list - should check assessments list endpoint", async () => {
-      const response = await fetch(`${API_URL}/api/assessment/list`, {
-        headers: { "Authorization": `Bearer ${testToken}` }
-      });
-      
-      // We expect 401 since our token is invalid, but endpoint should exist
-      console.log(`Assessment list endpoint status: ${response.status}`);
-      expect([200, 401, 403, 404]).toContain(response.status);
-    }, TEST_TIMEOUT);
-  });
-});
